# app/models.py
# generated by fastapi-codegen:
#   filename:  FF_API.yaml
#   timestamp: 2024-11-22T13:36:24+00:00

from __future__ import annotations

from datetime import date, datetime, time
from enum import Enum
from typing import List, Optional

from pydantic import AnyUrl, BaseModel, EmailStr, Field


class Error(BaseModel):
    code: str
    message: str


class Status(Enum):
    individual = 'individual'
    professional = 'professional'


class ActivityType(Enum):
    rental = 'rental'
    fishing_guide = 'fishing guide'


class LicenseType(Enum):
    coastal = 'coastal'
    river = 'river'


class BoatType(Enum):
    open = 'open'
    cabin = 'cabin'
    catamaran = 'catamaran'
    sailboat = 'sailboat'
    jet_ski = 'jet ski'
    canoe = 'canoe'


class EquipmentEnum(Enum):
    sonar = 'sonar'
    livewell = 'livewell'
    ladder = 'ladder'
    GPS = 'GPS'
    rod_holders = 'rod holders'
    VHF_radio = 'VHF radio'


class EngineType(Enum):
    diesel = 'diesel'
    gasoline = 'gasoline'
    none = 'none'


class Boat(BaseModel):
    id: Optional[str] = Field(None, description='Unique identifier of the boat')
    name: Optional[str] = Field(None, description='Name of the boat')
    description: Optional[str] = Field(None, description='Description of the boat')
    brand: Optional[str] = Field(None, description='Brand of the boat')
    manufactureYear: Optional[date] = Field(
        None, description='Year the boat was manufactured'
    )
    photoUrl: Optional[AnyUrl] = Field(None, description="URL of the boat's photo")
    licenseType: Optional[LicenseType] = Field(
        None, description='Required license type (coastal or river)'
    )
    boatType: Optional[BoatType] = Field(None, description='Type of boat')
    equipment: Optional[List[EquipmentEnum]] = Field(
        None,
        description='List of equipment on board (e.g., sonar, livewell, ladder, GPS, rod holders, VHF radio)',
    )
    depositAmount: Optional[float] = Field(
        None, description='Deposit amount (in euros)'
    )
    maxCapacity: Optional[int] = Field(None, description='Maximum number of passengers')
    numberOfBeds: Optional[int] = Field(None, description='Number of beds on the boat')
    homePort: Optional[str] = Field(None, description='Home port (city name)')
    latitude: Optional[float] = Field(None, description='Latitude of the home port')
    longitude: Optional[float] = Field(None, description='Longitude of the home port')
    engineType: Optional[EngineType] = Field(
        None, description='Type of engine (diesel, gasoline, or none)'
    )
    enginePower: Optional[int] = Field(None, description='Engine power (in horsepower)')


class TripType(Enum):
    daily = 'daily'
    recurring = 'recurring'


class RateType(Enum):
    total = 'total'
    per_person = 'per person'


class User(BaseModel):
    id: Optional[str] = Field(None, description='Unique identifier of the user')
    lastName: Optional[str] = None
    firstName: Optional[str] = None
    birthDate: Optional[date] = None
    email: Optional[EmailStr] = None
    boatLicense: Optional[str] = Field(
        None, description='Boat license number (8 digits)'
    )
    status: Optional[Status] = Field(
        None, description='Status of the user (individual or professional)'
    )
    companyName: Optional[str] = Field(
        None, description='Company name (empty if individual)'
    )
    activityType: Optional[ActivityType] = Field(
        None, description='Type of activity (rental or fishing guide)'
    )
    siretNumber: Optional[str] = Field(None, description='SIRET number')
    rcNumber: Optional[str] = Field(None, description='Commercial register number (RC)')
    boats: Optional[List[Boat]] = None
    trips: Optional[List[Trip]] = None
    reservations: Optional[List[Reservation]] = None
    log: Optional[Log] = None


class Trip(BaseModel):
    id: Optional[str] = Field(None, description='Unique identifier of the trip')
    title: Optional[str] = Field(None, description='Title of the trip')
    practicalInfo: Optional[str] = Field(
        None, description='Practical information about the trip'
    )
    tripType: Optional[TripType] = Field(
        None, description='Type of trip (daily or recurring)'
    )
    rateType: Optional[RateType] = Field(
        None, description='Type of rate (total or per person)'
    )
    startDates: Optional[List[datetime]] = Field(
        None, description='List of start dates for the trip'
    )
    endDates: Optional[List[datetime]] = Field(
        None, description='List of end dates for the trip'
    )
    departureTimes: Optional[List[time]] = Field(
        None, description='List of departure times'
    )
    endTimes: Optional[List[time]] = Field(None, description='List of end times')
    passengerCount: Optional[int] = Field(None, description='Number of passengers')
    price: Optional[float] = Field(None, description='Price of the trip')
    user: Optional[User] = None
    boat: Optional[Boat] = None


class Reservation(BaseModel):
    id: Optional[str] = Field(None, description='Unique identifier of the reservation')
    trip: Optional[Trip] = None
    date: Optional[datetime] = None
    reservedSeats: Optional[int] = None
    totalPrice: Optional[float] = None


class Log(BaseModel):
    id: Optional[str] = Field(None, description='Unique identifier of the log')
    pages: Optional[List[Page]] = None


class Page(BaseModel):
    id: Optional[str] = Field(None, description='Unique identifier of the page entry')
    fishName: Optional[str] = Field(None, description='Name of the fish')
    fishPhotoUrl: Optional[AnyUrl] = Field(None, description='URL of the fish photo')
    comment: Optional[str] = Field(None, description='Comment about the catch')
    length: Optional[float] = Field(
        None, description='Length of the fish in centimeters'
    )
    weight: Optional[float] = Field(None, description='Weight of the fish in kilograms')
    fishingSpot: Optional[str] = Field(None, description='Fishing location')
    fishingDate: Optional[date] = Field(None, description='Date of fishing')
    release: Optional[bool] = Field(
        None, description='Was the fish released? (true = yes, false = no)'
    )
    owner: Optional[User] = None


User.update_forward_refs()
Log.update_forward_refs()