# app/routers/log.py
# generated by fastapi-codegen:
#   filename:  FF_API.yaml
#   timestamp: 2024-11-22T13:36:24+00:00

from __future__ import annotations

from fastapi import APIRouter

from ..dependencies import *

router = APIRouter(tags=['Log'])


@router.get(
    '/v1/log/{user_id}',
    response_model=Log,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Log'],
)
def get_v1_log_user_id(user_id: str = Path(..., alias='userId')) -> Union[Log, Error]:
    """
    Get a user's log
    """
    pass


@router.post(
    '/v1/log/{user_id}',
    response_model=None,
    responses={
        '201': {'model': Log},
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '422': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Log'],
)
def post_v1_log_user_id(
    user_id: str = Path(..., alias='userId'), body: Log = ...
) -> Optional[Union[Log, Error]]:
    """
    Create a log for a user
    """
    pass


@router.put(
    '/v1/log/{user_id}/{page_id}',
    response_model=None,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Log'],
)
def put_v1_log_user_id_page_id(
    user_id: str = Path(..., alias='userId'),
    page_id: str = Path(..., alias='pageId'),
    body: Page = ...,
) -> Optional[Error]:
    """
    Edit a log page
    """
    pass


@router.delete(
    '/v1/log/{user_id}/{page_id}',
    response_model=None,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Log'],
)
def delete_v1_log_user_id_page_id(
    user_id: str = Path(..., alias='userId'), page_id: str = Path(..., alias='pageId')
) -> Optional[Error]:
    """
    Delete a log page
    """
    pass