# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Fisher Fans API",
    "description": "API for managing users, boats, trips, reservations, and logs.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.fisherfans.local/api",
      "description": "Secure local development server"
    }
  ],
  "paths": {
...
    "/v1/boats": {
      "get": {
        "summary": "Get the list of boats",
        "tags": [
          "Boats"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "Filter boats by user",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "y1",
            "in": "query",
            "description": "Minimum latitude for bounding box filter",
            "schema": {
              "type": "number",
              "format": "float"
            }
          },
          {
            "name": "y2",
            "in": "query",
            "description": "Maximum latitude for bounding box filter",
            "schema": {
              "type": "number",
              "format": "float"
            }
          },
          {
            "name": "x1",
            "in": "query",
            "description": "Minimum longitude for bounding box filter",
            "schema": {
              "type": "number",
              "format": "float"
            }
          },
          {
            "name": "x2",
            "in": "query",
            "description": "Maximum longitude for bounding box filter",
            "schema": {
              "type": "number",
              "format": "float"
            }
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "brand",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "boatType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "homePort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of boats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Boat"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "Create a new boat",
        "tags": [
          "Boats"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Boat"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Boat created"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/v1/boats/{boatId}": {
      "parameters": [
        {
          "name": "boatId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "summary": "Get a boat by ID",
        "tags": [
          "Boats"
        ],
        "responses": {
          "200": {
            "description": "Boat details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Boat"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "summary": "Edit a boat",
        "tags": [
          "Boats"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Boat"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Boat edited"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "Delete a boat",
        "tags": [
          "Boats"
        ],
        "responses": {
          "204": {
            "description": "Boat deleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/v1/log/{userId}": {
      "get": {
        "summary": "Get a user's log",
        "tags": [
          "Log"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The ID of the user whose log is requested"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Log successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Log"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "Create a log for a user",
        "tags": [
          "Log"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The ID of the user for whom the log is created"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Log"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Log successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Log"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/v1/log/{userId}/{pageId}": {
      "put": {
        "summary": "Edit a log page",
        "tags": [
          "Log"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Page"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Page successfully edited"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "Delete a log page",
        "tags": [
          "Log"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Page successfully deleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
  "components": {
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "Unprocessable entity",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the user"
          },
          "lastName": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "boatLicense": {
            "type": "string",
            "description": "Boat license number (8 digits)"
          },
          "status": {
            "type": "string",
            "description": "Status of the user (individual or professional)",
            "enum": [
              "individual",
              "professional"
            ]
          },
          "companyName": {
            "type": "string",
            "description": "Company name (empty if individual)"
          },
          "activityType": {
            "type": "string",
            "description": "Type of activity (rental or fishing guide)",
            "enum": [
              "rental",
              "fishing guide"
            ]
          },
          "siretNumber": {
            "type": "string",
            "description": "SIRET number"
          },
          "rcNumber": {
            "type": "string",
            "description": "Commercial register number (RC)"
          },
          "boats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Boat"
            }
          },
          "trips": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Trip"
            }
          },
          "reservations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reservation"
            }
          },
          "log": {
            "$ref": "#/components/schemas/Log"
          }
        }
      },
      "Boat": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the boat"
          },
          "name": {
            "type": "string",
            "description": "Name of the boat"
          },
          "description": {
            "type": "string",
            "description": "Description of the boat"
          },
          "brand": {
            "type": "string",
            "description": "Brand of the boat"
          },
          "manufactureYear": {
            "type": "string",
            "format": "date",
            "description": "Year the boat was manufactured"
          },
          "photoUrl": {
            "type": "string",
            "format": "uri",
            "description": "URL of the boat's photo"
          },
          "licenseType": {
            "type": "string",
            "description": "Required license type (coastal or river)",
            "enum": [
              "coastal",
              "river"
            ]
          },
          "boatType": {
            "type": "string",
            "description": "Type of boat",
            "enum": [
              "open",
              "cabin",
              "catamaran",
              "sailboat",
              "jet ski",
              "canoe"
            ]
          },
          "equipment": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "sonar",
                "livewell",
                "ladder",
                "GPS",
                "rod holders",
                "VHF radio"
              ]
            },
            "description": "List of equipment on board (e.g., sonar, livewell, ladder, GPS, rod holders, VHF radio)"
          },
          "depositAmount": {
            "type": "number",
            "format": "float",
            "description": "Deposit amount (in euros)"
          },
          "maxCapacity": {
            "type": "integer",
            "description": "Maximum number of passengers"
          },
          "numberOfBeds": {
            "type": "integer",
            "description": "Number of beds on the boat"
          },
          "homePort": {
            "type": "string",
            "description": "Home port (city name)"
          },
          "latitude": {
            "type": "number",
            "format": "float",
            "description": "Latitude of the home port"
          },
          "longitude": {
            "type": "number",
            "format": "float",
            "description": "Longitude of the home port"
          },
          "engineType": {
            "type": "string",
            "description": "Type of engine (diesel, gasoline, or none)",
            "enum": [
              "diesel",
              "gasoline",
              "none"
            ]
          },
          "enginePower": {
            "type": "integer",
            "description": "Engine power (in horsepower)"
          }
        }
      },
      "Trip": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the trip"
          },
          "title": {
            "type": "string",
            "description": "Title of the trip"
          },
          "practicalInfo": {
            "type": "string",
            "description": "Practical information about the trip"
          },
          "tripType": {
            "type": "string",
            "description": "Type of trip (daily or recurring)",
            "enum": [
              "daily",
              "recurring"
            ]
          },
          "rateType": {
            "type": "string",
            "description": "Type of rate (total or per person)",
            "enum": [
              "total",
              "per person"
            ]
          },
          "startDates": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "date-time"
            },
            "description": "List of start dates for the trip"
          },
          "endDates": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "date-time"
            },
            "description": "List of end dates for the trip"
          },
          "departureTimes": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "time"
            },
            "description": "List of departure times"
          },
          "endTimes": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "time"
            },
            "description": "List of end times"
          },
          "passengerCount": {
            "type": "integer",
            "description": "Number of passengers"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "Price of the trip"
          },
          "user": {
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              },
              {
                "description": "User organizing the trip"
              }
            ]
          },
          "boat": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Boat"
              },
              {
                "description": "Boat used for the trip"
              }
            ]
          }
        }
      },
      "Reservation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the reservation"
          },
          "trip": {
            "$ref": "#/components/schemas/Trip"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "reservedSeats": {
            "type": "integer"
          },
          "totalPrice": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "Log": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the log"
          },
          "pages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Page"
            }
          }
        }
      },
      "Page": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the page entry"
          },
          "fishName": {
            "type": "string",
            "description": "Name of the fish"
          },
          "fishPhotoUrl": {
            "type": "string",
            "format": "uri",
            "description": "URL of the fish photo"
          },
          "comment": {
            "type": "string",
            "description": "Comment about the catch"
          },
          "length": {
            "type": "number",
            "description": "Length of the fish in centimeters"
          },
          "weight": {
            "type": "number",
            "description": "Weight of the fish in kilograms"
          },
          "fishingSpot": {
            "type": "string",
            "description": "Fishing location"
          },
          "fishingDate": {
            "type": "string",
            "format": "date",
            "description": "Date of fishing"
          },
          "release": {
            "type": "boolean",
            "description": "Was the fish released? (true = yes, false = no)"
          },
          "owner": {
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              },
              {
                "description": "Owner of the fishing logbook"
              }
            ]
          }
        }
      }
    }
  },
...
}